int -> long in _phono3py.c under way

+++ c/fc3.h
@@ -36,34 +36,34 @@
 #define __fc3_H__
 
 void fc3_distribute_fc3(double *fc3,
-                        const int target,
-                        const int source,
-                        const int *atom_mapping,
+                        const long target,
+                        const long source,
+                        const long *atom_mapping,
                         const long num_atom,
                         const double *rot_cart);
 void fc3_rotate_delta_fc2(double (*fc3)[3][3][3],
                           PHPYCONST double (*delta_fc2s)[3][3],
                           const double *inv_U,
                           PHPYCONST double (*site_sym_cart)[3][3],
-                          const int *rot_map_syms,
+                          const long *rot_map_syms,
                           const long num_atom,
                           const long num_site_sym,
                           const long num_disp);
 void fc3_set_permutation_symmetry_fc3(double *fc3, const long num_atom);
 void fc3_set_permutation_symmetry_compact_fc3(double * fc3,
-                                              const int p2s[],
-                                              const int s2pp[],
-                                              const int nsym_list[],
-                                              const int perms[],
+                                              const long p2s[],
+                                              const long s2pp[],
+                                              const long nsym_list[],
+                                              const long perms[],
                                               const long n_satom,
                                               const long n_patom);
 void fc3_transpose_compact_fc3(double * fc3,
-                               const int p2s[],
-                               const int s2pp[],
-                               const int nsym_list[],
-                               const int perms[],
+                               const long p2s[],
+                               const long s2pp[],
+                               const long nsym_list[],
+                               const long perms[],
                                const long n_satom,
                                const long n_patom,
-                               const int t_type);
+                               const long t_type);
 
 #endif
